LayoutTests/imported/w3c/web-platform-tests/custom-elements/form-associated/form-reset-callback-expected.txt

default


PASS form.reset() should trigger formResetCallback
PASS form.reset(): formResetCallback is called synchronously at the end of form reset, with custom element reaction timing, *after* built-in form control reset algorithms run.
PASS Clicking a reset button invokes formResetCallback in a microtask
